Move remote app registry publish into daemon without a global helper bridge.
Remote publishing now lives in focused daemon files that call existing archive,
manifest, metadata, and command helpers directly. The appregistryremote package,
init-time registration, mirror archive types, and conversion glue are removed in
favor of explicit struct dependencies constructed at the CLI boundary for tests.
